emacs: change elisp function prefix from / to -
Function name prefix changed to - to follow the convention listed in
the official emacs manual.

"You should choose a short word to distinguish your program from other
Lisp programs. The names of all global symbols in your program, that
is the names of variables, constants, and functions, should begin with
that chosen prefix. Separate the prefix from the rest of the name with
a hyphen, ‘-’. This practice helps avoid name conflicts, since all
global variables in Emacs Lisp share the same name space, and all
functions share another name space33. Use two hyphens to separate
prefix and name if the symbol is not meant to be used by other
packages."