emacs: increased delay for autocompletions
All too often I get in the situation where I accidentally complete a
word when all I want is to create a new line.

I do still want to be able to use the <return> key for completions and
so the decision has been made to just increase the delay to something
sensible. If I want completions right away I can just invoke it with C-M-i.