emacs(org): config restructure
I have now figured out what is needed for being able to both use the
latest version of Org, and be able to configure it in config.org.

I leave only the installation part inside of init.el and make sure to
use `:ensure nil` inside the configuraiton to ensure that the
configuration does not try to install the package again, if missing.