emacs(org): split journelly template
Splits template into one with generated metadata and one that is
plain. This makes it so that I still can add notes, even though I
don't have access to the library required for metadata generation on
non-macOS systems.

**Possible future modification:**
Considering replacing the code for generating location and weather
data with a more generic one that works on multiple platforms.