emacs(theme): auto dark config refactor
Uses global constants for specifying dark and light theme. The consts
are then referenced in the configuration. This makes it so that we
only need to make a change in one place if we decide to use different
themes for dark/light.

Current light theme: modus-operandi.
Current dark theme:  ef-night.