shell: bashrc.d files sourced, not executed
Since the files in .bashrc.d are sourced and not executed, we don't
need the shebangs. Practically speaking we don't need the file
extensions either, but we add them to get syntax highlighting etc. out
of the box.

If possible in their current implementations, we use the .sh
extension. Right now, the .bash files have bash-specific logic that
would need to be ported for them to be sh compatible. This is for
another time.